extract multisig code into a mod
